Static analysis upgrade testing (#269)
Use static analysis to do upgrade testing.

A blunt hammer to provide some basic upgrade testing, if touching a class or field that is suspected of causing trouble with upgrades the test will fail. Also adding some annotations and other ways to allow the test to skip over things determined to be safe.